requirements.txt
email-validator>=2.2.0
flask>=3.1.1
flask-sqlalchemy>=3.1.1
gunicorn>=23.0.0
numpy>=2.3.1
opencv-python>=4.11.0.86
pillow>=11.3.0
psycopg2-binary>=2.9.10
scikit-learn>=1.7.0
tensorflow>=2.14.0
tensorflow-cpu>=2.14.0
werkzeug>=3.1.3
